Initial commit of the project

+++ src/main/java/com/Springboot/Kirana_Inventory_Management/Repo/SaleItemRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.Springboot.Kirana_Inventory_Management.Entity.SaleItem;
 
 @Repository
-public interface SaleItemRepository extends JpaRepository<SaleItem, Long> {
+public interface SaleItemRepository extends JpaRepository<SaleItem, Integer> {
 }
